requirements.txt
absl-py==2.1.0
attrs==23.2.0
cffi==1.16.0
contourpy==1.2.0
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
flatbuffers==23.5.26
fonttools==4.47.2
kiwisolver==1.4.5
matplotlib==3.8.2
mediapipe==0.10.9
numpy==1.26.3
opencv-contrib-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.2
pillow==10.2.0
protobuf==3.20.3
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.8.2
six==1.16.0
sounddevice==0.4.6
